Revert "Finalizado funcionalidade da list view e carregando cards dinamicamente."
This reverts commit 0f231c274adb75cbd7133a3c3655802c1d90c233.

+++ app/src/main/java/com/example/projetopdm/adapter/RvAdapter.java
@@ -40,7 +40,7 @@ public class RvAdapter extends RecyclerView.Adapter<RvAdapter.PersonViewHolder> 
     @Override
     public void onBindViewHolder(@NonNull PersonViewHolder holder, int i) {
         holder.Titulo.setText(viagens.get(i).getNome());
-        holder.data.setText(Long.toString(viagens.get(i).getQtde_dias()));
+        holder.data.setText(Integer.toString(viagens.get(i).getQtde_dias()));
         holder.valor_total.setText(Float.toString(viagens.get(i).getTotal_viagem()));
     }
